# src/main/resources/swagger.yaml
openapi: 3.0.3
info:
  title: Watch the Chamber - BFF API
  description: >
    Backend For Frontend API between the mobile app and the Brazilian Chamber of Deputies public API.
    Handles authentication, user management, followed politicians, and notifications.
  version: 1.0.0
  contact:
    name: Raphael Coelho
    email: contato@raphaelcoelho.dev
servers:
  - url: http://localhost:8080/api/v1
    description: Local environment
  - url: https://api.deolhonacamara.com/api/v1
    description: Production
...
paths:
...
  /api/v1/politicians/{id}/expenses:
    get:
      tags: [ Expenses ]
      summary: Get expenses for a specific politician
      operationId: getPoliticianExpenses
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: query
          name: page
          schema: { type: integer, default: 0 }
        - in: query
          name: size
          schema: { type: integer, default: 20 }
        - in: query
          name: year
          schema: { type: integer }
          description: Filter by year
        - in: query
          name: month
          schema: { type: integer }
          description: Filter by month (1-12)
      responses:
        '200':
          description: Paginated list of expenses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseResponseDTO'
...
  /api/v1/politicians/{id}/speeches:
    get:
      tags: [ Speeches ]
      summary: Get speeches for a specific politician
      operationId: getPoliticianSpeeches
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: query
          name: page
          schema: { type: integer, default: 0 }
        - in: query
          name: size
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Paginated list of speeches
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpeechResponseDTO'
...
  /api/v1/sync/all:
    post:
      tags: [ Sync ]
      summary: Trigger full synchronization of all features
      operationId: syncAll
      description: Starts an asynchronous synchronization process for all features (politicians, expenses, votes, speeches, propositions, and presence). Returns immediately without waiting for completion.
      responses:
        '202':
          description: Synchronization started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncResponse'
        '500':
          description: Error starting synchronization
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ExpenseDto:
      type: object
      properties:
        id:
          type: integer
        politicianId:
          type: integer
        year:
          type: integer
        month:
          type: integer
        expenseType:
          type: string
        supplier:
          type: string
        documentValue:
          type: string
        netValue:
          type: string
        glosaValue:
          type: string
        documentDate:
          type: string
          format: date
        documentUrl:
          type: string
    ExpenseResponseDTO:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ExpenseDto'
        detail:
          type: object
          properties:
            year:
              type: integer
            month:
              type: integer
            totalExpenses:
              type: string
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
        sizePage:
          type: integer
...
    SpeechDto:
      type: object
      properties:
        startDateTime:
          type: string
        endDateTime:
          type: string
        titleEvent:
          type: string
        endDateTimeEvent:
          type: string
        startDateTimeEvent:
          type: string
        summary:
          type: string
        speechType:
          type: string
        transcription:
          type: string
        eventUri:
          type: string
        audioUrl:
          type: string
        textUrl:
          type: string
        videoUrl:
          type: string
    SpeechResponseDTO:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/SpeechDto'
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
        sizePage:
          type: integer
...
    SyncResponse:
      type: object
      properties:
        message:
          type: string
          example: Synchronization started successfully
        status:
          type: string
          example: ACCEPTED