# target/generated-sources/openapi/src/main/java/net/coelho/deolhonacamara/api/SpeechesApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.13.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package net.coelho.deolhonacamara.api;

import net.coelho.deolhonacamara.api.model.SpeechResponseDTO;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-21T10:53:37.437611200-03:00[America/Fortaleza]", comments = "Generator version: 7.13.0")
@Validated
@Tag(name = "Speeches", description = "Politician speeches in the chamber")
public interface SpeechesApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /api/v1/politicians/{id}/speeches : Get speeches for a specific politician
     *
     * @param id  (required)
     * @param page  (optional, default to 0)
     * @param size  (optional, default to 20)
     * @return Paginated list of speeches (status code 200)
     */
    @Operation(
        operationId = "getPoliticianSpeeches",
        summary = "Get speeches for a specific politician",
        tags = { "Speeches" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Paginated list of speeches", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SpeechResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/politicians/{id}/speeches",
        produces = { "application/json" }
    )
    
    default ResponseEntity<SpeechResponseDTO> getPoliticianSpeeches(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id,
        @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Integer page,
        @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "20") Integer size
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"total\" : 0, \"data\" : [ { \"summary\" : \"summary\", \"eventUri\" : \"eventUri\", \"audioUrl\" : \"audioUrl\", \"startDateTime\" : \"startDateTime\", \"transcription\" : \"transcription\", \"videoUrl\" : \"videoUrl\", \"endDateTimeEvent\" : \"endDateTimeEvent\", \"speechType\" : \"speechType\", \"startDateTimeEvent\" : \"startDateTimeEvent\", \"endDateTime\" : \"endDateTime\", \"titleEvent\" : \"titleEvent\", \"textUrl\" : \"textUrl\" }, { \"summary\" : \"summary\", \"eventUri\" : \"eventUri\", \"audioUrl\" : \"audioUrl\", \"startDateTime\" : \"startDateTime\", \"transcription\" : \"transcription\", \"videoUrl\" : \"videoUrl\", \"endDateTimeEvent\" : \"endDateTimeEvent\", \"speechType\" : \"speechType\", \"startDateTimeEvent\" : \"startDateTimeEvent\", \"endDateTime\" : \"endDateTime\", \"titleEvent\" : \"titleEvent\", \"textUrl\" : \"textUrl\" } ], \"totalPages\" : 1, \"sizePage\" : 5, \"page\" : 6 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}